Home

Run.jl

Run.docs

Run.prepare_docs

Run.prepare_test

Run.test

Run.test — Function.

Run.test(path="test"; prepare, fast, compiled_modules, strict, precompile)

Run $path/runtests.jl after activating $path/Project.toml.

Keyword Arguments

prepare::Bool = true: Call Run.prepare_test if true (default).

fast::Bool = false: Try to run it faster (more precisely, pass --compile=min option to Julia subprocess.)

compiled_modules::Union{Bool, Nothing} = nothing: Use --compiled-modules=yes (--compiled-modules=no) option if true (false). If false, it also skips precompilation in the preparation phase.

strict::Bool = true: Do not include the default environment in the load path (more precisely, set the environment variable JULIA_LOAD_PATH=@).

Other keywords are passed to Run.prepare_test.

source

Run.docs — Function.

Run.docs(path="docs"; prepare, fast, compiled_modules, strict, precompile)

See Run.test.

source

Run.prepare_test — Function.

Run.prepare_test(path="test"; precompile)

Instantiate $path/Project.toml. It also devs the project in the parent directory of path into $path/Project.toml if $path/Manifest.toml does not exist.

Keyword Arguments

precompile::Bool = true: Precompile the project if true (default).

source

Run.prepare_docs — Function.

Run.prepare_docs(path="docs"; precompile)

See Run.prepare_test.

source